Void equipment fixes (#15)
* Add gloves to void set, fix void multipliers, add elite void magic damage bonus

* Change magic damage to fully display elite void bonus

* Change int to float and add DecimalFormat

* Merge overloaded buildStatLabel methods

+++ src/main/java/com/duckblade/osrs/dpscalc/model/ItemStats.java
@@ -28,7 +28,7 @@ public class ItemStats
 	@SerializedName("rstr")
 	private int strengthRanged;
 	@SerializedName("mdmg")
-	private int strengthMagic;
+	private float strengthMagic;
 
 	@SerializedName(value = "speed", alternate = {"aspeed"})
 	private int speed;
